+++ src/app/(headerless)/login/hooks/useLogin.ts
@@ -1,0 +1,11 @@
+import { useAuth } from '@/hooks/useAuth';
+import { useRouter } from 'next/navigation';
+import { useEffect } from 'react';
+
+export default function useLogin() {
+  const { auth } = useAuth();
+  const router = useRouter();
+  useEffect(() => {
+    if (auth !== null) router.push('/home');
+  }, [auth, router]);
+}
