+++ src/app/(headerless)/oauth/page.tsx
@@ -1,35 +1,8 @@
 'use client';
 
-import { useAuth } from '@/hooks/useAuth';
-import { auth } from '@/stores/authStore';
-import { useRouter } from 'next/navigation';
-import { useEffect } from 'react';
+import useOauth from './hooks/useOauth';
 
 export default function OauthPage() {
-  //TODO url로 직접 접근 막기
-  const { setAuth } = useAuth();
-  const router = useRouter();
-
-  useEffect(() => {
-    const code = new URL(window.location.href).searchParams.get('code');
-    (async () => {
-      await fetch(
-        `${process.env.NEXT_PUBLIC_TEST_API_END_POINT}/mock/login?code=${code}`,
-        {
-          method: 'POST',
-        },
-      )
-        .then((res) => res.json())
-        .then((data) => {
-          const auth: auth = {
-            accessToken: data.accessToken,
-            refreshToken: data.refreshToken,
-          };
-          setAuth(auth);
-        });
-
-      router.push('/home');
-    })();
-  }, [setAuth, router]);
-  return <>OauthPage,hello👋</>;
+  useOauth();
+  return;
 }
